+++ packages/fields/src/actions/trash-post.tsx
@@ -107,10 +107,10 @@ const trashPost: Action< PostWithPermissions > = {
 									);
 								} else {
 									successMessage = sprintf(
-										/* translators: The number of items. */
+										/* translators: %d: The number of items. */
 										_n(
-											'%s item moved to the trash.',
-											'%s items moved to the trash.',
+											'%d item moved to the trash.',
+											'%d items moved to the trash.',
 											items.length
 										),
 										items.length
@@ -138,7 +138,7 @@ const trashPost: Action< PostWithPermissions > = {
 									}
 									// If we were trying to delete multiple items.
 								} else {
-									const errorMessages = new Set();
+									const errorMessages = new Set< string >();
 									const failedPromises = promiseResult.filter(
 										( { status } ) => status === 'rejected'
 									);